refactoring for new pipeline and adding logging functionality
