+++ Dockerfile
@@ -8,15 +8,15 @@ ENV ASPNETCORE_URLS=http://*:3000
 FROM mcr.microsoft.com/dotnet/core/sdk:3.1 AS build
 WORKDIR /src
 
-COPY src/DotNetCore_Template/DotNetCore_Template.csproj .
-RUN dotnet restore DotNetCore_Template.csproj
+COPY src/App/App.csproj .
+RUN dotnet restore App.csproj
 COPY . .
 WORKDIR /src/.
-RUN dotnet build "DotNetCore_Template.csproj" -c Release -o /app/build
+RUN dotnet build "src/App/App.csproj" -c Release -o /app/build
 
 ###########################################################################
 FROM build AS publish
-RUN dotnet publish "src/DotNetCore_Template/DotNetCore_Template.csproj" -c Release -o /app/publish
+RUN dotnet publish "src/App/App.csproj" -c Release -o /app/publish
 
 ##########################################################################
 LABEL maintainer="%CUSTOM_PLUGIN_CREATOR_USERNAME%" \
@@ -36,4 +36,4 @@ RUN echo "%CUSTOM_PLUGIN_SERVICE_NAME%: $COMMIT_SHA" >> ./commit.sha
 
 COPY --from=publish /app/publish .
 
-CMD ["dotnet", "DotNetCore_Template.dll"]
+CMD ["dotnet", "App.dll"]
