--- DotNetCore_Template.sln
@@ -1,23 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetCore_Template", "src\DotNetCore_Template\DotNetCore_Template.csproj", "{23D222A8-CE9A-4833-9D33-2AE9E6457265}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetCore_Template.Tests", "src\DotNetCore_Template.Tests\DotNetCore_Template.Tests.csproj", "{9EE88146-3F6F-4C4A-904F-7EAB94A0A1D4}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{23D222A8-CE9A-4833-9D33-2AE9E6457265}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{23D222A8-CE9A-4833-9D33-2AE9E6457265}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{23D222A8-CE9A-4833-9D33-2AE9E6457265}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{23D222A8-CE9A-4833-9D33-2AE9E6457265}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9EE88146-3F6F-4C4A-904F-7EAB94A0A1D4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9EE88146-3F6F-4C4A-904F-7EAB94A0A1D4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9EE88146-3F6F-4C4A-904F-7EAB94A0A1D4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9EE88146-3F6F-4C4A-904F-7EAB94A0A1D4}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
