+++ src/App/Properties/AssemblyInfo.cs
@@ -1,0 +1,1 @@
+[assembly: log4net.Config.XmlConfigurator(ConfigFile = "log4net.config")]
