+++ src/App/appsettings.Development.json
@@ -2,8 +2,8 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
+      "Microsoft": "Error",
+      "Microsoft.Hosting.Lifetime": "None"
     }
   },
   "USERID_HEADER_KEY": "userid",
@@ -11,6 +11,5 @@
   "GROUPS_HEADER_KEY": "usergroups",
   "CLIENTTYPE_HEADER_KEY": "clienttype",
   "BACKOFFICE_HEADER_KEY": "isbackoffice",
-  "MICROSERVICE_GATEWAY_SERVICE_NAME": "microservice-gateway",
-  "CRUD_PATH": "crud"
+  "MICROSERVICE_GATEWAY_SERVICE_NAME": "microservice-gateway"
 }
