+++ src/App/log4net.config
@@ -1,0 +1,11 @@
+<log4net>
+    <root>
+      <level value="ALL" />
+      <appender-ref ref="console" />
+    </root>
+    <appender name="console" type="log4net.Appender.ConsoleAppender">
+      <layout type="log4net.Layout.PatternLayout">
+        <conversionPattern value="%message%newline" />
+      </layout>
+    </appender>
+  </log4net>
